pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "medical-image-segmentation"
version = "0.4.0"
description = "Medical AI: YOLOv8 Detection + VLM Interpretation + RAG Evidence"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.32.0",
    "python-multipart==0.0.18",
    "python-dotenv>=1.0.0",
    "ultralytics==8.4.14",
    "opencv-python-headless==4.10.0.84",
    "langchain==0.3.27",
    "langchain-core==0.3.83",
    "langchain-community==0.3.27",
    "langchain-chroma==0.2.6",
    "langchain-huggingface==0.3.1",
    "langchain-ollama==0.2.3",
    "langchain-openai==0.3.35",
    "langchain-text-splitters==0.3.11",
    "langgraph>=0.2.0",
    "sentence-transformers>=3.0.0",
    "pymupdf>=1.24.0",
    "openai>=1.40.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "prometheus-client>=0.20.0",
    "psutil>=6.0.0",
    "structlog>=24.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "httpx>=0.27.0", "ruff>=0.8.0", "pre-commit>=3.0"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.setuptools.packages.find]
where = ["."]
include = ["api*", "core*", "db*", "eval*", "preprocessing*", "rag*", "scripts*", "training*", "vlm*"]
exclude = ["docker*", "example_test*", "tests*", "ui*"]
namespaces = true
